LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-activation/activation-traverse-not-in-entries-expected.txt



FAIL navigation.activation - traverse from a non-contiguous same-origin url assert_equals: expected 1 but got 0
